# src/main/java/com/probestack/forgestudio/design/api/EventsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.AnnotateAuditEventRequest;
import com.probestack.forgestudio.design.model.AuditEvent;
import com.probestack.forgestudio.design.model.AuditEventListResponse;
import com.probestack.forgestudio.design.model.CreateAuditEventRequest;
import org.springframework.format.annotation.DateTimeFormat;
import com.probestack.forgestudio.design.model.ErrorResponse;
import java.time.OffsetDateTime;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-14T07:29:28.385377570Z[GMT]")@Validated
@Tag(name = "Events", description = "Operations on audit events")
public interface EventsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * PUT /events/{eventId} : Annotate an audit event
     * Adds a compliance annotation to an existing event. The original event payload remains immutable.
     *
     * @param eventId Unique audit event identifier (UUID) (required)
     * @param annotateAuditEventRequest  (required)
     * @return Audit event annotated successfully (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "annotateAuditEvent",
        summary = "Annotate an audit event",
        description = "Adds a compliance annotation to an existing event. The original event payload remains immutable.",
        tags = { "Events" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Audit event annotated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuditEvent.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/events/{eventId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AuditEvent> annotateAuditEvent(
        @Parameter(name = "eventId", description = "Unique audit event identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("eventId") String eventId,
        @Parameter(name = "AnnotateAuditEventRequest", description = "", required = true) @Valid @RequestBody AnnotateAuditEventRequest annotateAuditEventRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"annotation\" : \"Reviewed during Q3 audit\", \"eventId\" : \"a1u2d3t4\", \"resourceId\" : \"INV-2026-00317\", \"afterValue\" : { \"key\" : \"\" }, \"occurredAt\" : \"2026-08-12T10:00:00Z\", \"beforeValue\" : { \"key\" : \"\" }, \"actorRole\" : \"ORG_ADMIN\", \"ipAddress\" : \"203.0.113.42\", \"action\" : \"UPDATE\", \"actorEmail\" : \"finance.admin@example.com\", \"resourceType\" : \"Invoice\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * POST /events : Record an audit event
     * Appends a new event to the immutable log. The &#x60;eventId&#x60; is generated by the server. Events cannot be modified once written.
     *
     * @param createAuditEventRequest  (required)
     * @return Audit event recorded successfully (status code 201)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "createAuditEvent",
        summary = "Record an audit event",
        description = "Appends a new event to the immutable log. The `eventId` is generated by the server. Events cannot be modified once written.",
        tags = { "Events" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Audit event recorded successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuditEvent.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/events",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AuditEvent> createAuditEvent(
        @Parameter(name = "CreateAuditEventRequest", description = "", required = true) @Valid @RequestBody CreateAuditEventRequest createAuditEventRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"annotation\" : \"Reviewed during Q3 audit\", \"eventId\" : \"a1u2d3t4\", \"resourceId\" : \"INV-2026-00317\", \"afterValue\" : { \"key\" : \"\" }, \"occurredAt\" : \"2026-08-12T10:00:00Z\", \"beforeValue\" : { \"key\" : \"\" }, \"actorRole\" : \"ORG_ADMIN\", \"ipAddress\" : \"203.0.113.42\", \"action\" : \"UPDATE\", \"actorEmail\" : \"finance.admin@example.com\", \"resourceType\" : \"Invoice\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /events/{eventId} : Get an audit event by ID
     * Retrieves a single audit event including its before and after value snapshots.
     *
     * @param eventId Unique audit event identifier (UUID) (required)
     * @return Audit event found (status code 200)
     *         or Audit event not found (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "getAuditEventById",
        summary = "Get an audit event by ID",
        description = "Retrieves a single audit event including its before and after value snapshots.",
        tags = { "Events" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Audit event found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuditEvent.class))
            }),
            @ApiResponse(responseCode = "404", description = "Audit event not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/events/{eventId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuditEvent> getAuditEventById(
        @Parameter(name = "eventId", description = "Unique audit event identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("eventId") String eventId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"annotation\" : \"Reviewed during Q3 audit\", \"eventId\" : \"a1u2d3t4\", \"resourceId\" : \"INV-2026-00317\", \"afterValue\" : { \"key\" : \"\" }, \"occurredAt\" : \"2026-08-12T10:00:00Z\", \"beforeValue\" : { \"key\" : \"\" }, \"actorRole\" : \"ORG_ADMIN\", \"ipAddress\" : \"203.0.113.42\", \"action\" : \"UPDATE\", \"actorEmail\" : \"finance.admin@example.com\", \"resourceType\" : \"Invoice\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /events : List audit events
     * Returns a paginated list of audit events, optionally filtered by actor, action, or time window.
     *
     * @param actorEmail Filter by the user who performed the action (optional)
     * @param action Filter by action type (optional)
     * @param resourceType Filter by the type of resource acted upon (e.g., \&quot;Invoice\&quot;) (optional)
     * @param fromDate Return events at or after this timestamp (optional)
     * @param toDate Return events at or before this timestamp (optional)
     * @param page Page number (starts at 1) (optional, default to 1)
     * @param size Number of items per page (optional, default to 20)
     * @return Successful response (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "listAuditEvents",
        summary = "List audit events",
        description = "Returns a paginated list of audit events, optionally filtered by actor, action, or time window.",
        tags = { "Events" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuditEventListResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/events",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuditEventListResponse> listAuditEvents(
        @jakarta.validation.constraints.Email @Parameter(name = "actorEmail", description = "Filter by the user who performed the action", in = ParameterIn.QUERY) @Valid @RequestParam(value = "actorEmail", required = false) String actorEmail,
        @Parameter(name = "action", description = "Filter by action type", in = ParameterIn.QUERY) @Valid @RequestParam(value = "action", required = false) String action,
        @Parameter(name = "resourceType", description = "Filter by the type of resource acted upon (e.g., \"Invoice\")", in = ParameterIn.QUERY) @Valid @RequestParam(value = "resourceType", required = false) String resourceType,
        @Parameter(name = "fromDate", description = "Return events at or after this timestamp", in = ParameterIn.QUERY) @Valid @RequestParam(value = "fromDate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime fromDate,
        @Parameter(name = "toDate", description = "Return events at or before this timestamp", in = ParameterIn.QUERY) @Valid @RequestParam(value = "toDate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime toDate,
        @Min(1) @Parameter(name = "page", description = "Page number (starts at 1)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "1") Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "Number of items per page", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"size\" : 20, \"totalPages\" : 1241, \"page\" : 1, \"content\" : [ { \"annotation\" : \"Reviewed during Q3 audit\", \"eventId\" : \"a1u2d3t4\", \"resourceId\" : \"INV-2026-00317\", \"afterValue\" : { \"key\" : \"\" }, \"occurredAt\" : \"2026-08-12T10:00:00Z\", \"beforeValue\" : { \"key\" : \"\" }, \"actorRole\" : \"ORG_ADMIN\", \"ipAddress\" : \"203.0.113.42\", \"action\" : \"UPDATE\", \"actorEmail\" : \"finance.admin@example.com\", \"resourceType\" : \"Invoice\" }, { \"annotation\" : \"Reviewed during Q3 audit\", \"eventId\" : \"a1u2d3t4\", \"resourceId\" : \"INV-2026-00317\", \"afterValue\" : { \"key\" : \"\" }, \"occurredAt\" : \"2026-08-12T10:00:00Z\", \"beforeValue\" : { \"key\" : \"\" }, \"actorRole\" : \"ORG_ADMIN\", \"ipAddress\" : \"203.0.113.42\", \"action\" : \"UPDATE\", \"actorEmail\" : \"finance.admin@example.com\", \"resourceType\" : \"Invoice\" } ], \"totalElements\" : 24810 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * DELETE /events/{eventId} : Purge an audit event
     * Removes an event that has passed its retention period. Events within retention cannot be purged.
     *
     * @param eventId Unique audit event identifier (UUID) (required)
     * @return Audit event purged successfully (no content returned) (status code 204)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "purgeAuditEvent",
        summary = "Purge an audit event",
        description = "Removes an event that has passed its retention period. Events within retention cannot be purged.",
        tags = { "Events" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Audit event purged successfully (no content returned)"),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/events/{eventId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> purgeAuditEvent(
        @Parameter(name = "eventId", description = "Unique audit event identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("eventId") String eventId
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }

}